api_python/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==3.2.2
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyJWT==2.12.1
PyMySQL==1.1.2
pytest==9.0.3
python-dotenv==1.2.2
python-multipart==0.0.26
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
